Create Showing the Login Form
In order to show the login form in case a user has not logged in, and hence has not created a session, we use another if and else statement around the #wrapper div and script tags in our original code. In the opposite case, this will hide the login form and show the chat box if the user is logged in and has created a session.